[Add] 영어 키보드 기능 구현
